patched error while creating only binary systems

+++ scripts_python/core_program_create_def.py
@@ -100,27 +100,29 @@ for gamma_folder in gamma_folder_list:
 		DEF_part[cell] = os.path.join(dir_script,"references", f"DEFINITIONS_{cell}.txt")
 	R_part = {"part1": R1_np, "part2": R2_np}
 	
-	dir_fuente = {"part1": os.path.join(dir_inicial,"sim_part1"),"part2": os.path.join(os.getcwd(),"part2")}
-	for label in ["part1", "part2"]:
-		os.chdir(dir_fuente[label])
-		for label_struc in cell_part:
-			os.chdir(os.path.join(dir_fuente[label],label_struc))
-			dir_fuente_part = dir_fuente[label]
-			shutil.copy(DEF_part[label_struc], os.path.join(os.getcwd(),"DEFINITIONS.txt"))
-			DEF = os.path.join(os.getcwd(), "DEFINITIONS.txt")
-			lines = read_DEF(DEF)
-			size_index = None
-			for i, line in enumerate(lines):
-				if line.strip() == "!particle semiaxis x y z in nm":
-					size_index = i + 1
-			for n in np.arange(0,k_part[label_struc]):
-				lines[size_index + n] = f"{R_part[label]} {R_part[label]} {R_part[label]}\n"
-			write_DEF(DEF, lines) 
+	if os.path.exists(os.path.join(dir_fuente,"part2")):
+		print(cell_part)
+		dir_fuente = {"part1": os.path.join(dir_inicial,"sim_part1"),"part2": os.path.join(os.getcwd(),"part2")}
+		for label in ["part1", "part2"]:
+			os.chdir(dir_fuente[label])
+			for label_struc in cell_part:
+				os.chdir(os.path.join(dir_fuente[label],label_struc))
+				dir_fuente_part = dir_fuente[label]
+				shutil.copy(DEF_part[label_struc], os.path.join(os.getcwd(),"DEFINITIONS.txt"))
+				DEF = os.path.join(os.getcwd(), "DEFINITIONS.txt")
+				lines = read_DEF(DEF)
+				size_index = None
+				for i, line in enumerate(lines):
+					if line.strip() == "!particle semiaxis x y z in nm":
+						size_index = i + 1
+				for n in np.arange(0,k_part[label_struc]):
+					lines[size_index + n] = f"{R_part[label]} {R_part[label]} {R_part[label]}\n"
+				write_DEF(DEF, lines) 
 
-			DEF = os.path.join(os.getcwd(), "DEFINITIONS.txt")
-			R_np = extract_R_part(DEF)
-			aL = float(run_command(f'python3 {dir_script}/references/aL_min_{label_struc}.py {R_np}'))
-			process_principal_part(DEF, delta_part[label_struc], aL, tosubmit, dir_fuente[label])
+				DEF = os.path.join(os.getcwd(), "DEFINITIONS.txt")
+				R_np = extract_R_part(DEF)
+				aL = float(run_command(f'python3 {dir_script}/references/aL_min_{label_struc}.py {R_np}'))
+				process_principal_part(DEF, delta_part[label_struc], aL, tosubmit, dir_fuente[label])
 
 import pandas as pd
 folder_ref = [os.path.join(dir_inicial,"sim_part1/binary_ref")]
